projects/project3/examples/하/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
